\name{SampleSize-class}
\Rdversion{1.1}
\docType{class}
\alias{SampleSize-class}

\title{Class "SampleSize"}
\description{A class used internally only that contains an integer (the
  estimated sample size).
}

\section{Objects from the Class}{
  These objects are only used internally.
}

\section{Slots}{
  \describe{
    \item{\code{estimate}:}{Object of class \code{"integer"} }
  }
}

\section{Methods}{
No methods defined with class "SampleSize" in the signature.
}

\author{\email{thomas.fabbro@unibas.ch}}

\examples{
showClass("SampleSize")
}

\keyword{internal}
\keyword{classes}
