\name{tex-methods}
\docType{methods}
\alias{tex-methods}
\alias{tex,power,character-method}
\alias{tex}
\title{Methods for Function tex}
\description{
  Methods for function \code{tex}
}
\usage{\S4method{tex}{power,character}(x, type, ...)}
\arguments{
  \item{x}{The object of class \code{power} used for extraction}
  \item{type}{Currently available:
    \itemize{
      \item{\code{drop}}
      \item{\code{nRec}}
      \item{\code{nEval}}
      \item{\code{n.iter}}
      \item{\code{power}}
      \item{\code{sampling}}
	\item{\code{theta}}
    }}
  \item{...}{Not used so far}
  }
\section{Methods}{
\describe{

\item{}{describe this method here}
}}
\keyword{methods}

