\name{tex}
\docType{methods}
\alias{tex}
\title{Methods for Function tex}
\description{
  Methods for function \code{tex}
}
\usage{tex(x, type, ...)}
\arguments{
  \item{x}{The object of class \code{power} used for extraction}
  \item{type}{Currently available:
    \itemize{
      \item{\code{drop}}
      \item{\code{nRec}}
      \item{\code{nEval}}
      \item{\code{n.iter}}
      \item{\code{power}}
      \item{\code{sampling}}
	\item{\code{theta}}
    }}
  \item{...}{Not used so far}
  }
\section{Methods}{
\describe{This methods prepare strings that can directly be used for
  including information from objects of \linkS4class{power} into Sweave reports.
  }}
\keyword{methods}

